add OpenAI embedding support
