+++ docs/docs/reference/config.md
@@ -78,6 +78,8 @@ See [ML Inference overview](../features/inference/overview.md) and [RAG Search](
 
 ### Embedding models
 
+#### ONNX models
+
 Example of a full configuration:
 
 ```yaml
@@ -106,6 +108,29 @@ Fields:
 * `batch_size`: *optional*, *int*, default `32`. Computing embeddings is a highly parallel task, and doing it in big chunks is much more effective than one by one. For CPUs there are usually no gains of batch sizes beyong 32, but on GPUs you can go up to 1024.
 * `pooling`: *optional*, `cls`/`mean`, default auto. Which pooling method use to compute sentence embeddings. This is model specific and Nixiesearch tries to guess it automatically. See the [list of supported embeddings models](../features/inference/embeddings/sbert.md) to know if it can be detected automatically. If your model is not on the list, consult the model doc on its pooling method.
 * `normalize`: *optional*, *bool*, default `true`. Should embeddings be L2-normalized? With normalized embeddings it becomes possible to use a faster dot-product based aKNN search.
+
+#### OpenAI models
+
+Example of a full configuration:
+
+```yaml
+inference:
+  embedding:
+    <model-name>:
+      provider: openai
+      model: text-embedding-3-small
+      timeout: 2000ms
+      endpoint: "https://api.openai.com/"
+      dimensions: null
+      batch_size: 32
+```
+
+Parameters:
+
+* **timeout**: optional, duration, default 2s. External APIs might be slow sometimes.
+* **retry**: optional, string, default "https://api.openai.com/". You can use alternative API or EU-specific endpoint.
+* **dimensions**: optional, int, default empty. For [matryoshka](https://huggingface.co/blog/matryoshka) models, how many dimensions to return.
+* **batch_size**: optional, int, default 32. Batch size for calls with many documents.
 
 ### LLM completion models
