+++ src/test/resources/config/compat.yml
@@ -3,9 +3,6 @@ inference:
     text:
       provider: onnx
       model: nixiesearch/e5-small-v2-onnx
-      prompt:
-        query: "query: "
-        doc: "passage: "
 
 
 schema:
