+++ src/test/resources/config/msmarco.yml
@@ -3,9 +3,6 @@ inference:
     text:
       provider: onnx
       model: nixiesearch/e5-small-v2-onnx
-      prompt:
-        query: "query: "
-        doc: "passage: "
 
 store:
   url: file:///tmp/msmarco/
